Tova branch

+++ client/src/components/ProfileForm.js
@@ -1,22 +1,20 @@
-import React, { useState } from "react";
+import React, { useContext } from "react";
 import { Formik } from 'formik';
 import * as yup from 'yup'
 import { Container } from '@mui/material';
 import { useNavigate } from "react-router-dom";
+import { AppContext } from "../context/Context";
 
-function ProfileForm({ user, setUser }) {
-    // const [profile, setProfile] = useState(true)
+function ProfileForm({ setUser }) {
     const navigate = useNavigate();
-
+    const useAppContext = () => useContext(AppContext);
+    const { user } = useAppContext();
 
     const profileSchema = yup.object().shape({
         firstName: yup.string().min(1, 'First name too short!').max(15, 'First name too long!'),
         lastName: yup.string().min(1, 'Last name too short!').max(15, 'Last name too long!'),
-        email: yup.string().email("Invalid email address"),//.min(1, 'Email too short!').max(25, 'Email too long!'),
+        email: yup.string().email("Invalid email address"),
         phoneNumber: yup.string().min(10, 'Phone number too short!').max(17, 'Phone number too long!'),
-        // username: yup.string().min(5, 'Username too short!').max(15, 'Username too long!'),
-        // password: yup.string().min(5, 'Password too short!').max(15, 'Password too long!'),
-        // passwordConfirmation: yup.string().oneOf([yup.ref('password')], 'Passwords must match!'),
         zipCode: yup.number().integer().min(10000, 'Invalid zip code').max(99999, 'Invalid zip code!'),
         interests: yup.string().required('Interest is required!'),
         skills: yup.string().required('Skills are required!'),
@@ -46,6 +44,9 @@ function ProfileForm({ user, setUser }) {
     }
 
     const handleAccountDelete =(values) => {
+        if (!window.confirm("Are you sure you want to delete your account?")){
+            return;
+        }
         const endpoint = `/volunteer/${user.id}`
         fetch(endpoint, {
             method: 'DELETE',
@@ -64,24 +65,20 @@ function ProfileForm({ user, setUser }) {
         });
     }
 
-
     const initialValues = {
-        firstName: '',
-        lastName: '',
-        email: '',
-        phoneNumber: '',
-        // username: '',
-        // password: '',
-        // passwordConfirmation: '',
-        zipCode: '',
-        interests: '',
-        skills: '',
-        hoursWanted: ''
+        firstName: user === null || user === undefined ? '' : user.first_name,
+        lastName: user === null || user === undefined ? '' : user.last_name,
+        email: user === null || user === undefined ? '' : user.email,
+        phoneNumber: user === null || user === undefined ? '' : user.phone_number,
+        zipCode: user === null || user === undefined ? '' : user.zipcode,
+        interests: user === null || user === undefined ? '' : user.interests,
+        skills: user === null || user === undefined ? '' : user.skills,
+        hoursWanted: user === null || user === undefined ? '' : user.hours_wanted
     }
 
     return (
         <Container className="profile-container">
-            <Formik
+            <Formik enableReinitialize
                 initialValues={initialValues}
                 validationSchema={profileSchema}
                 onSubmit={handleFormSubmit}
@@ -121,40 +118,6 @@ function ProfileForm({ user, setUser }) {
                                 onChange={handleChange}
                             />
                         </div>
-                        {/* <div className="right-column">
-                        <label htmlFor='username'>Username:</label>
-                        <input
-                            id='username'
-                            name='username'
-                            type='username'
-                            placeholder='Username'
-                            value={values.username}
-                            onChange={handleChange}
-                        />
-                    </div> */}
-                        {/* {/* </div> */}
-                        {/* <div className="left-column">
-                        <label htmlFor='password'>Password:</label>
-                        <input
-                            id='password'
-                            name='password'
-                            type='password'
-                            placeholder='Password'
-                            value={values.password}
-                            onChange={handleChange}
-                        />
-                    </div> */}
-                        {/* <div className="right-column">
-                        <label htmlFor='passwordConfirmation'>Password Confirmation:</label>
-                        <input
-                            id='passwordConfirmation'
-                            name='passwordConfirmation'
-                            type='password'
-                            placeholder='Password Confirmation'
-                            value={values.passwordConfirmation}
-                            onChange={handleChange}
-                        />
-                    </div>  */}
                         <div className="right-column">
                             <label htmlFor='zipCode'>Zip Code:</label>
                             <input
@@ -188,17 +151,6 @@ function ProfileForm({ user, setUser }) {
                                 onChange={handleChange}
                             />
                         </div>
-                        <div className="right-column">
-                            <label htmlFor='interests'>Interests:</label>
-                            <input
-                                id='interests'
-                                name='interests'
-                                type='interests'
-                                placeholder='Interests'
-                                value={values.interests}
-                                onChange={handleChange}
-                            />
-                        </div>
                         <div className="left-column">
                             <label htmlFor='skills'>Skills:</label>
                             <input
@@ -210,6 +162,18 @@ function ProfileForm({ user, setUser }) {
                                 onChange={handleChange}
                             />
                         </div>
+                        <div className="right-column">
+                            <label htmlFor='interests'>Interests:</label>
+                            <input
+                                id='interests'
+                                name='interests'
+                                type='interests'
+                                placeholder='Interests'
+                                value={values.interests}
+                                onChange={handleChange}
+                            />
+                        </div>
+                        
                         <div className="button-container-left">
                             <button type='submit'>Submit</button>
                         </div>
